+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 ### Adjustments
 - remove ```Save and Close``` of fuzzing dialog, as closing means saving here anyway
 
+### New features
+- make application __font size__ configurable in ``` Options ``` -> ``` View ```
+
 ## v2.5.1 (25/11/2018)
 ### Bugfixes
 - improve stability when capturing with SoundCard